index.js added to routes folder
