+++ middleware/validation.js
@@ -75,7 +75,6 @@ const validateUser = celebrate({
     password: Joi.string().required().messages({
       "string.empty": 'The "password" field must be filled in',
     }),
-
     name: Joi.string().min(2).max(30).messages({
       "string.min": 'The minimum length of the "name" field is 2',
       "string.max": 'The maximum length of the "name" field is 30',
